#32 Refactor Go package outside of 'Yi'
