+++ common/src/main/kotlin/codes/nibby/go/AnnotationType.kt
@@ -1,4 +1,4 @@
-package codes.nibby.yi.go
+package codes.nibby.go
 
 enum class AnnotationType {
