+++ common/src/main/kotlin/codes/nibby/go/GameRulesException.kt
@@ -1,4 +1,4 @@
-package codes.nibby.yi.go
+package codes.nibby.go
 
 /**
  * Thrown when an attempted action is a violation of game rules.
