+++ common/src/main/kotlin/codes/nibby/go/GoGamePosition.kt
@@ -1,4 +1,4 @@
-package codes.nibby.yi.go
+package codes.nibby.go
 
 class GoGamePosition(private val boardWidth: Int, boardHeight: Int) {
