+++ common/src/main/kotlin/codes/nibby/go/GoMoveHelper.kt
@@ -1,4 +1,4 @@
-package codes.nibby.yi.go
+package codes.nibby.go
 
 import java.util.*
 import kotlin.collections.HashSet
