+++ common/src/main/kotlin/codes/nibby/go/GoStoneColor.kt
@@ -1,4 +1,4 @@
-package codes.nibby.yi.go
+package codes.nibby.go
 
 import java.lang.IllegalStateException
