+++ common/src/main/kotlin/codes/nibby/go/MoveSequence.kt
@@ -1,4 +1,4 @@
-package codes.nibby.yi.go
+package codes.nibby.go
 
 /**
  * A manager class that enables a sequence of moves to be played while checking each move is played successfully.
