+++ common/src/main/kotlin/codes/nibby/go/MoveTree.kt
@@ -1,4 +1,4 @@
-package codes.nibby.yi.go
+package codes.nibby.go
 
 open class MoveTree<NodeData> constructor(rootNode: MoveNode<NodeData>) {
