+++ common/src/main/kotlin/codes/nibby/go/rules/ChineseRulesHandler.kt
@@ -1,4 +1,4 @@
-package codes.nibby.yi.go.rules
+package codes.nibby.go.rules
 
 class ChineseRulesHandler : GoGameRulesHandler() {
