+++ editor/src/main/java/codes/nibby/yi/editor/Main.java
@@ -2,9 +2,9 @@ package codes.nibby.yi.editor;
 
 import codes.nibby.yi.editor.gui.board.GameBoard;
 import codes.nibby.yi.editor.settings.Settings;
-import codes.nibby.yi.go.Annotation;
-import codes.nibby.yi.go.GoGameRules;
-import codes.nibby.yi.go.GoGameModel;
+import codes.nibby.go.Annotation;
+import codes.nibby.go.GoGameRules;
+import codes.nibby.go.GoGameModel;
 import javafx.application.Application;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
