+++ editor/src/main/java/codes/nibby/yi/editor/gui/board/PresetStoneStyle.java
@@ -1,6 +1,6 @@
 package codes.nibby.yi.editor.gui.board;
 
-import codes.nibby.yi.go.GoStoneColor;
+import codes.nibby.go.GoStoneColor;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.effect.BlurType;
 import javafx.scene.effect.DropShadow;
